pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssh-capsule"
version = "0.1.0"
description = "Bootstrap reproducible SSH environments on any server with one command"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "ssh-capsule contributors"}]
keywords = ["ssh", "server", "provisioning", "devops", "automation", "reproducible"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "paramiko>=3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
ssh-capsule = "ssh_capsule.cli:cli"

[project.urls]
Homepage = "https://github.com/ssh-capsule/ssh-capsule"
Issues = "https://github.com/ssh-capsule/ssh-capsule/issues"

[tool.setuptools.package-data]
ssh_capsule = ["templates/*.yaml"]
